Cisco Mech Replay n times and suppress exceptions
There are two parts to this changeset.  First allow user to configure
a finite number of replay retries by the Cisco Mechanism replay code.
Second to suppress exceptions from nexus driver calls when the
Cisco Mechanism replay is enabled.  Details of each follows.

For the first part 'limit replay retries', the Cisco Mechanism Driver
has a new configuration variable 'switch_replay_count' in local.conf
file beneath 'ml2_cisco'.  This configures the max number of replay
attempts.  By default, it is set to 3 and is dependent on the config
variable 'switch_heartbeat_time' > 0.  If replay fails the configured
number of times, it will stop attempting to send to Nexus device.
If any configuration received from higher layer when switch state is
disabled will merely be accumulated in the local db and not sent to
the Nexus device.  However, if there is new configuration activity,
this will cause the replay code to reset it's retry counter to zero
and we will again start over to attempt to replay configuration to the
remote for the configured number of time.

For the 2nd part of this change set, exceptions will no longer
make their way up to higher layer code when Cisco Mechanism Replay
is enabled.  This is to avoid collisions between two replay
drivers.  So when the local replay is not running, exceptions
will be allowed up to higher layer replay code.

Change-Id: Ic59086021d76630df489df7c25836dc3d58bf65c
Closes-bug:  #1422738