Don't raise exception for hosts not configured
Remove calls to raise an exception when a host name is not found under
the cisco nexus switch configuration section. Under ML2, other MD's
may process these hosts.

This exception was also being raised from the bind_port() method when
a host name was not found. Code changed to not process event for hosts
not configured.

Change-Id: I1accb767fb75854034a28f7047b6e52449fd245b
Closes-Bug: 1424867