ASR1k: DB update error configuring >= 70 routers with gateway set
The issue here is that the ASR1k plugins uses the row used by in the agent DB
to send the list of routers ids (in the configuration cell). Since the schema
for this cell in the table has a maximum size of 4096, then around 70 routers
we exceed this number.

A solution would be to separate this row into a different table. However, it
seems that the original purpose of this field is to send the hosting_device_id
and the total number of routers. Hence I'm fixing it by sending the intended
number (there's one thing to mention that in the ASR1k plugin case, the
hosting_device_id is one of the two ASR1k routers).

Example of the new format of configurations for 8 tenant routers:
'configurations': {
       'total routers': 9,
       'non_responding_hosting_devices': [],
       'hosting_devices': {u'aaaaaaaa-1234-5678-9876-zzzzzzzzzzzz': {'routers': 1},
                           u'b29fbcbd-3234-4a8f-a593-175c0ca5f8e7': {'routers': 3},
                           u'd0de1e16-6e50-44e8-806c-73b5b8152f66': {'routers': 5}},
       'total floating_ips': 0,
       'total ex_gw_ports': 8,
       'total interfaces': 48}}}

This addresses bug DE329:
https://us1.rallydev.com/#/31072017263d/detail/defect/38522499203

Change-Id: Ib97186be19ddd28277997cd840927fc5a65f0c4a