ASR1K: Fullsync errors after ctlr reboot / plugin restart
The ASR1K plugin gets stuck in the fullsync tyring to delete old config from
ASR1K after DB corruption (due to for example controller reboot or plugin restart).
This is because some initializations of router list or dict are not done propertly.
These need to be created as {} or dict() but, right now, they are created as None,
which can cause problems:
One of the problems happens when there's RPC exception communicating with the
ASR1K routers, a TypeError will happen because of of this initialized as None instead
of dict() (or just {})

This addresses DE334 bug:
https://us1.rallydev.com/#/31072017263d/detail/defect/38589447976

Change-Id: Ic0c56846b39418f16e7c1ef5ba501372fdee580b