ASR1K: Floating IP status updates on dis/associating a fix IP
Currently the floating IP gets created and its status is set to
ACTIVE and never changes. This causes Jenkins cases to fail.

This commit covers the changes needed in the ASR1K plugin to update
the status field of a Floating IP (it gets created to DOWN and then
the cfg_agent backend notify neutron to update the status to ACTIVE
when the floating IP is associated or th a fixed IP or DOWN when
it gets disassociated).

Example of a floating IP not associated:

+---------------------+--------------------------------------+
| Field               | Value                                |
+---------------------+--------------------------------------+
| fixed_ip_address    |                                      |
| floating_ip_address | 100.100.8.228                        |
| floating_network_id | 6e9273b8-67d0-4c4c-8c73-df9e8bef0dfc |
| id                  | fccf6dcd-9559-4f7a-b468-ef766ceddad0 |
| port_id             |                                      |
| router_id           |                                      |
| status              | DOWN                                 |
| tenant_id           | b94778e325794b7398f3cacdeb8f6064     |
+---------------------+--------------------------------------+

Note: The next patch will cover the fixes for the floating IP UT

Change-Id: I6f60e35b1313f5f23a72424d4513ab151b3d43a9
Closes-Bug:#1479500