ML2 cisco_nexus MD: Temp Fix for large scale bug
As described in DE588,
"With neutron multiworkers configured, there is a potential race
condition issue where some of the VLANs will not be configured on one
or more N9k switches.

/etc/neutron/neutron.conf
------------------
api_workers=3
rpc_workers=3"

Workaround is to allow the vlan create command to be sent down to a
switch under most event conditions. Long term fix will be to introduce
a new column in the port binding DB table that indicates the true state
of the entry/row.

Change-Id: If57dd35d2a055b1e35e6b79dbd085f7add3825b6