ASR1K: Low performance of plugin in scale conditions
In fullsync case, it takes more than 40 mins to read the router
interfaces on the neutron DB and the plugin fails with an exception
due to RPC timeout of 2 minutes. This commit addresses this issue
and it reduces drastically the time on neutron DB queries.

This is done by simplifying the queries. For example, the one
that access the interface in fullsync, we don't filter by router_id
since at this moment the plugin assumes all routers are ASR1K.
Also it looks like accessing DB is very expensive. So, it's
preferable to query less, make a dict with the results and
then access the dict.

This addressed DE684: ASR1k Plugin: replay issue with scale

Change-Id: I9158e17ed542c8ca53ea3c82887d15366653213e