ASR1K: no standby needs v2 to avoid netconf errors
Netconf errors were observed on the ASR1K while the plugin was configuring
interfaces.  The sequence of configuration the plugin does first attempts
to delete the standby on the interface prior to configuring it.
If no standby is configured then the ASR assumes the interface is a
V1 standby interface with a group number limit of 0-255.   Our plugin
uses V2 group numbers in the range of 0-4095 and whenever a group number
larger than 255 is used netconf would error during the delete.

This fix updates the snippets to use standby V2 when attempting to
delete the standby.

Change-Id: I34f5e84a6f28cc70b413361bfcea028c964d4d74