ASR1K: Wipe-out config if conf_sync fails due to timeout
If netconf returns timeout error, we are now clearing the config of the ASRs.
This commit fixes this by raising the Exception so that we don't erase
the router_ids table and proceed with trying to read/clear/cfg the ASR.
Then in the next periodical update, we'll re-try to sync the config to the ASRs.

Change-Id: I37a7933cf34656f3de14874a80a5e63c82166ad4