ML2 Cisco Nexus MD: VLAN/trunk config not created on all switches
In the following topology/configuration:
- two nexus switches
- three neutron-server controllers (HA)
- dhcp_agents_per_network=3
- two port-channels being configured per nexus switch

a race condition was observed that caused one switch missing a vlan global
configuration and one of the port-channels not being trunked with the vlan.
The 2nd switch was often not programmed at all.

This fix removes the code that attempts to verify (and then program) if a
vlan/trunk interface was programmed on the switch.

(Double commit of bug from networking-cisco/master to
staging/junoplus cisco repo.)

Change-Id: If384cfabf4dd72939abe5565233b44413b5ffee1
Closes-Bug: #1512463